% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbnl.xrf.data-package.R
\docType{package}
\name{lbnl.xrf.data-package}
\alias{lbnl.xrf.data}
\alias{lbnl.xrf.data-package}
\title{lbnl.xrf.data: LBNL XRF Data}
\description{
Data contained in this package were generated by the Lawrence Berkeley National Laboratory's Nuclear Archaeology Program, in collaboration with others, using X-ray fluorescence spectrometry.
}
\author{
\strong{Maintainer}: Matthew Boulanger \email{mboulanger@smu.edu} (\href{https://orcid.org/0000-0003-3384-7884}{ORCID})

}
\keyword{internal}
