% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbnl_xrf.R
\docType{data}
\name{lbnl_xrf}
\alias{lbnl_xrf}
\title{Geochemical data for Andes obsidian sources and artifacts}
\format{
A tibble with 510 rows and 8 variables:
\describe{
\item{ANID}{chr Analytical ID unique to each specimen analyzed.}
\item{Site}{chr Locality, site, or source from which the piece was collected.}
\item{\code{Fe}}{dbl Abundance of iron (Fe) in percent.}
\item{\code{Rb}}{dbl Abundance of rubidium (Rb) in ppm.}
\item{\code{Sr}}{dbl Abundance of strontium (Sr) in ppm.}
\item{\code{Zr}}{dbl Abundance of zirconium (Zr) in ppm.}
\item{\code{Ba}}{dbl Abundance of barium (Ba) in ppm.}
\item{\code{Ce}}{dbl Abundance of cerium (Ce) in ppm.}
}
}
\source{
\url{https://doi.org/10.1080/00776297.2022.2029157}
}
\usage{
lbnl_xrf
}
\description{
Data were produced at the LBNL Nuclear Archaeology Program, in collaboration with R. Burger.
}
\keyword{datasets}
